requirements.txt
yt-dlp>=2023.3.4
moviepy>=1.0.3
mutagen>=1.46.0
pillow>=9.4.0
requests>=2.28.2